// src/types.ts
export interface BubbleColorThemePreset {
  id: string;
  name: string;
  userBg: string;
  userBorder: string;
  userText: string;
  assistantBg: string;
  assistantBorder: string;
  assistantText: string;
}